Moved some files to new dir /core/
