+++ src/core/YINIVisitor.ts
@@ -1,6 +1,5 @@
 import assert from 'assert'
-import { isDebug } from './config/env'
-import { ErrorDataHandler } from './ErrorDataHandler'
+import { isDebug } from '../config/env'
 import {
     Boolean_literalContext,
     ElementContext,
@@ -21,21 +20,22 @@ import {
     Terminal_lineContext,
     ValueContext,
     YiniContext,
-} from './grammar/YiniParser.js'
-import YiniParserVisitor from './grammar/YiniParserVisitor'
-import { extractYiniLine } from './parsers/extractSignificantYiniLine'
-import parseBooleanLiteral from './parsers/parseBoolean'
-import parseNullLiteral from './parsers/parseNull'
-import parseNumberLiteral from './parsers/parseNumber'
-import parseSectionHeader from './parsers/parseSectionHeader'
-import parseStringLiteral from './parsers/parseString'
+} from '../grammar/YiniParser.js'
+import YiniParserVisitor from '../grammar/YiniParserVisitor'
+import { extractYiniLine } from '../parsers/extractSignificantYiniLine'
+import parseBooleanLiteral from '../parsers/parseBoolean'
+import parseNullLiteral from '../parsers/parseNull'
+import parseNumberLiteral from '../parsers/parseNumber'
+import parseSectionHeader from '../parsers/parseSectionHeader'
+import parseStringLiteral from '../parsers/parseString'
+import { debugPrint, printObject } from '../utils/system'
+import { ErrorDataHandler } from './ErrorDataHandler'
 import {
     IChainContainer,
     ISectionResult,
     TSyntaxTree,
     TSyntaxTreeContainer,
 } from './types'
-import { debugPrint, printObject } from './utils/system'
 
 interface YiniDocument {
     // sections: Record<string, any>
