+++ src/data-extractors/parseSectionHeader.ts
@@ -2,10 +2,10 @@ import { isDebug } from '../config/env'
 import { ErrorDataHandler } from '../core/ErrorDataHandler'
 import { TSectionHeaderType } from '../core/types'
 import { SectionContext } from '../grammar/YiniParser'
+import extractHeaderParts from '../parsers/extractHeaderParts'
+import { extractYiniLine } from '../parsers/extractSignificantYiniLine'
 import { isAlpha } from '../utils/string'
 import { debugPrint } from '../utils/system'
-import extractHeaderParts from './extractHeaderParts'
-import { extractYiniLine } from './extractSignificantYiniLine'
 
 const SECTION_MARKER1 = '^'
 const SECTION_MARKER2 = '~'
