+++ src/parsers/extractHeaderParts.ts
@@ -1,7 +1,7 @@
 import { isDebug } from '../config/env'
-import { ErrorDataHandler } from '../ErrorDataHandler'
+import { ErrorDataHandler } from '../core/ErrorDataHandler'
+import { TSectionHeaderType } from '../core/types'
 import { SectionContext } from '../grammar/YiniParser'
-import { TSectionHeaderType } from '../types'
 import { trimBackticks } from '../utils/string'
 import { debugPrint } from '../utils/system'
 import { extractYiniLine } from './extractSignificantYiniLine'
